Force parsing HTTP request as JSON object
When sending an HTTP request from the Google Cloud Platform Scheduler,
the online interface does not allow you to specify the mimetype for your
request. By default, it does not use application/json, so let's ignore
mimetype for now and always try to parse requests as json since that's
the kind of data we are expecting.