Send weather notifications from target function
Update the target function to send weather notifications. A weather
notification will be 'all clear' if the current forecast does not predict over
40% chance of rain in the next 24hrs. If the forecast does predict > 40% chance
of rain in at least 1hr in the next 24hrs, then the weather notification
contains a rain warning. This is a pretty rough estimate, but we'll use this
algorithm for now.

Remove the 'hello world' code that was used for testing Google Cloud Platform.